Update API Account1vs1, Fix field of report and notification, update …

+++ ThinkTank.Data/Entities/Contest.cs
@@ -19,7 +19,7 @@ namespace ThinkTank.Data.Entities
         public bool? Status { get; set; }
         public int? CoinBetting { get; set; }
         public int? GameId { get; set; }
-        public int? PlayTime { get; set; }
+        public decimal? PlayTime { get; set; }
 
         public virtual Game? Game { get; set; }
         public virtual ICollection<AccountInContest> AccountInContests { get; set; }
