+++ ThinkTank.Service/DTO/Request/CreateAndUpdateContestRequest.cs
@@ -16,7 +16,7 @@ namespace ThinkTank.Service.DTO.Request
         public int CoinBetting { get; set; }
         public int GameId { get; set; }
         [Range(1, int.MaxValue, ErrorMessage = "Only positive number allowed")]
-        public int PlayTime { get; set; }
+        public decimal PlayTime { get; set; }
         public List<CreateAssetOfContestRequest> Assets { get; set; }
     }
 }
